move skins location and fix character skin card show
1.将skins文件移动到src的skins目录下;
2.修复大厅--武将图--首图重复时的显示问题;
3.将dev模式下的getCharacterImage方法恢复到原始版本。
