+++ src/ui/platforms/desktop/src/audio_loader/dev_audio_loader.ts
@@ -1,6 +1,6 @@
 import { CharacterGender } from 'core/characters/character';
 import { AudioLoader } from './audio_loader';
-import { CharacterSkinInfo } from '../image_loader/skins';
+import { CharacterSkinInfo } from 'skins/skins';
 
 const remoteRoot: string = 'http://doublebit.gitee.io/pictest/backup_remote';
